requirements.txt
fastapi
uvicorn
psutil
python-multipart
jinja2
python-dotenv